requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
bcrypt==4.1.2
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.4
cryptography==42.0.7
cymem==2.0.8
cytoolz==0.12.3
defusedxml==0.8.0rc2
Django==5.0.4
django-allauth==0.63.0
django-cors-headers==4.3.1
django-filter==24.2
django-oauth-toolkit==2.3.0
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
fastembed==0.2.7
filelock==3.14.0
flatbuffers==24.3.25
floret==0.10.5
fsspec==2024.3.1
grpcio==1.63.0
grpcio-tools==1.63.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
intel-openmp==2021.4.0
jellyfish==1.0.3
Jinja2==3.1.4
joblib==1.4.2
jwcrypto==1.5.6
langcodes==3.4.0
language_data==1.2.0
loguru==0.7.2
marisa-trie==1.1.1
Markdown==3.6
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
onnx==1.16.0
onnxruntime==1.17.3
packaging==24.0
pillow==10.3.0
portalocker==2.8.2
preshed==3.0.9
protobuf==5.26.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
PyMySQL==1.1.0
pyphen==0.15.0
pyreadline3==3.4.1
python3-openid==3.2.0
pywin32==306
PyYAML==6.0.1
qdrant-client==1.9.1
regex==2024.5.10
requests==2.31.0
requests-oauthlib==2.0.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
smart-open==6.4.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.0
srsly==2.4.8
sympy==1.12
tbb==2021.12.0
textacy==0.13.0
thinc==8.2.3
threadpoolctl==3.5.0
tokenizers==0.19.1
toolz==0.12.1
torch==2.3.0
tqdm==4.66.4
transformers==4.40.2
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
win32-setctime==1.1.0
